% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.context.R
\name{get.context}
\alias{get.context}
\title{Get Context}
\usage{
get.context(file, start)
}
\arguments{
\item{file}{Reference files directory}

\item{start}{The position of the mutation gene}
}
\value{
The trinucleotide context.
}
\description{
Gets the 5' and 3' neighboring bases to the mutated base
}
\examples{
example.ref.dir <- paste0(
	path.package("SeqKat"),
	"/extdata/test/ref/"
	);
get.context(file.path(example.ref.dir, 'chr4.fa'), c(1582933, 1611781))
}
\author{
Fouad Yousif

Fan Fan
}
