% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.exprobntcx.R
\name{get.exprobntcx}
\alias{get.exprobntcx}
\title{get.exprobntcx}
\usage{
get.exprobntcx(somatic, ref.dir, trinucleotide.count.file)
}
\arguments{
\item{somatic}{Data frame of somatic variants}

\item{ref.dir}{Path to a directory containing the reference genome.}

\item{trinucleotide.count.file}{A tab seprarated file containing a count of all trinucleotides present in the reference genome. This can be generated with the get.trinucleotide.counts() function in this package.}
}
\description{
Gets the expected probability for each trinucleotide and total number of tcx
}
\examples{
load(
	paste0(
		path.package("SeqKat"),
		"/extdata/test/somatic.rda"
		)
	);

trinucleotide.count.file <- paste0(
	path.package("SeqKat"),
	"/extdata/tn_count.txt"
	);

example.ref.dir <- paste0(
	path.package("SeqKat"),
	"/extdata/test/ref/"
	);

get.exprobntcx(somatic, example.ref.dir, trinucleotide.count.file)
}
\author{
Fan Fan

Fouad Yousif
}
