% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nucleotide.chunk.counts.R
\name{get.nucleotide.chunk.counts}
\alias{get.nucleotide.chunk.counts}
\title{Get Nucleotide Chunk Counts}
\usage{
get.nucleotide.chunk.counts(key, chr, upstream = 1, downstream = 1,
  start = 1, end = -1)
}
\arguments{
\item{key}{List of specify trinucleotides to count}

\item{chr}{Chromosome}

\item{upstream}{Length upstream to read}

\item{downstream}{Length downstream to read}

\item{start}{Starting position}

\item{end}{Ending position}
}
\description{
Obtain counts for all possible trinucleotides within a specified genomic region
}
\examples{
example.ref.dir <- paste0(
	path.package("SeqKat"),
	"/extdata/test/ref/"
	);

bases.raw <- c('A','C','G','T','N');
tri.types.raw <- c(
	outer(
		c(outer(bases.raw, bases.raw, function(x, y) paste0(x,y))),
		bases.raw, function(x, y) paste0(x,y))
	);
tri.types.raw <- sort(tri.types.raw);
get.nucleotide.chunk.counts(
	tri.types.raw,
	file.path(example.ref.dir, 'chr4.fa'),
	upstream = 1,
	downstream = 1,
	start = 1,
	end = -1
	);
}
\author{
Fouad Yousif
}
