% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pair.R
\name{get.pair}
\alias{get.pair}
\title{Get Pair}
\usage{
get.pair(x)
}
\arguments{
\item{x}{asdf}
}
\description{
Generates the reverse compliment of a nucleotide sequence
}
\details{
Reverses and compliments the bases of the input string. Bases must be (A, C, G, T, or N).
}
\examples{
get.pair("GATTACA")
}
\author{
Fouad Yousif
}
