% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tn.R
\name{get.tn}
\alias{get.tn}
\title{Get Trinucleotides}
\usage{
get.tn(chr, start.bp, end.bp, ref.dir)
}
\arguments{
\item{chr}{Chromosome}

\item{start.bp}{Starting position}

\item{end.bp}{Ending position}

\item{ref.dir}{Path to a directory containing the reference genome.}
}
\description{
Count the frequencies of 32 trinucleotide in a region respectively
}
\examples{
example.ref.dir <- paste0(
	path.package("SeqKat"),
	"/extdata/test/ref/"
	);
get.tn(chr=4, start.bp=1, end.bp=-1, example.ref.dir)
}
\author{
Fan Fan
}
