% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.toptn.R
\name{get.toptn}
\alias{get.toptn}
\title{Get Top Trinucleotides}
\usage{
get.toptn(somatic.subset, chr, start.bp, end.bp, ref.dir)
}
\arguments{
\item{somatic.subset}{Data frame of somatic variants subset for a specific chromosome}

\item{chr}{Chromosome}

\item{start.bp}{Starting position}

\item{end.bp}{Ending position}

\item{ref.dir}{Path to a directory containing the reference genome.}
}
\description{
Generate a tri-nucleotide summary for each sliding window
}
\examples{
\dontrun{
get.toptn(somatic.subset, chr, start.bp, end.bp, ref.dir)
}

}
\author{
Fan Fan

Fouad Yousif
}
