% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trinucleotide.counts.R
\name{get.trinucleotide.counts}
\alias{get.trinucleotide.counts}
\title{Get Trinucleotide Counts}
\usage{
get.trinucleotide.counts(ref.dir, ref.name, output.dir)
}
\arguments{
\item{ref.dir}{Path to a directory containing the reference genome.}

\item{ref.name}{Name of the reference genome being used (i.e. hg19, GRCh38, etc)}

\item{output.dir}{Path to a directory where output will be created.}
}
\description{
Aggregates the total counts of each possible trinucleotide.
}
\examples{
\dontrun{
get.trinucleotide.counts(ref.dir, "hg19", tempdir());
}
}
\author{
Fan Fan

Fouad Yousif
}
